% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchTtest.R
\name{batchTtest}
\alias{batchTtest}
\title{qPCR Batch T-test}
\usage{
batchTtest(df, ctr = "pcas13d", hsk = "cdk2ap2", outputPath)
}
\arguments{
\item{df}{Dataframe imported from importFiles.}

\item{ctr}{Defines a character string with the name of your control group used to perform T-test in comparison to. Default is 'pcas13d'}

\item{hsk}{Defines a character string with the housekeeping gene used in the experiment. Default is "cdk2ap2"}

\item{outputPath}{Path to save files, plots and the report}
}
\value{
Dataframe with T-test results from all pairwise comparisons.
}
\description{
This function performs T-test between your ctr group and all other
}
