% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta_CT.R
\name{calculate_DCT}
\alias{calculate_DCT}
\title{Normalize HKG}
\usage{
calculate_DCT(df, hkg, sample_col = "Sample", gene_col = "Target")
}
\arguments{
\item{df}{tibble}

\item{hkg}{vector of genes to normalize to}

\item{sample_col}{column name of samples}

\item{gene_col}{column name of genes}
}
\value{
df
}
\description{
Normalize to a selection of (housekeeping)genes
}
