% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta_delta_CT.R
\name{calculate_DDCT}
\alias{calculate_DDCT}
\title{Calculate ddct}
\usage{
calculate_DDCT(df = qpcr, gene_col, sample_col, var_col, control)
}
\arguments{
\item{df}{tibble}

\item{gene_col}{column name of genes}

\item{sample_col}{column name of samples}

\item{var_col}{column name of variables to normalize against your control}

\item{control}{name of variable to use as control}
}
\value{
tibble
}
\description{
Calculate the delta delta ct
}
