% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{calculate_t.test}
\alias{calculate_t.test}
\title{Statistics}
\usage{
calculate_t.test(df = qpcr, x, y)
}
\arguments{
\item{df}{qpcr dataframe}

\item{x}{x variable for hypothesis test}

\item{y}{y variable for hypothesis test}
}
\value{
tidy data frame
}
\description{
Calculate t.test
}
