% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultTechCtrs.R
\docType{data}
\name{defaultTechCtrs}
\alias{defaultTechCtrs}
\title{Default technical controls}
\format{
An object of class \code{character} of length 11.
}
\usage{
defaultTechCtrs
}
\description{
This is the default technical controls used in the package.
}
\keyword{datasets}
