% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaDeltaCtCalc.R
\name{deltaDeltaCtCalc}
\alias{deltaDeltaCtCalc}
\title{qPCR delta delta Ct calculation}
\usage{
deltaDeltaCtCalc(df, hsk = "cdk2ap2", ctr = "pcas13d", outputPath)
}
\arguments{
\item{df}{Dataframe imported from importFiles.}

\item{hsk}{Defines a character string with the name of your housekeeping gene used to normalize the qPCR data. Default is 'cdk2ap2'}

\item{ctr}{Defines a character string with the name of your control group used to normalize delta-delta-Cts. Default is 'pcas13d'}

\item{outputPath}{Path to save files, plots and the report}
}
\value{
Outputs a dataframe with delta delta Ct values based on each target.
}
\description{
This function calculates the delta delta Ct values based on the specified housekeeping gene.
}
