% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagOutliers.R
\name{flagOutliers}
\alias{flagOutliers}
\title{qPCR Flag Technical Outliers}
\usage{
flagOutliers(
  df,
  techCtrs = defaultTechCtrs,
  outputPath,
  sdCutOff = 0.2,
  assayInfo,
  sampleInfo
)
}
\arguments{
\item{df}{Dataframe imported from importFiles.}

\item{techCtrs}{Character string for technical controls (e.g., water, RT-). Default is within defaultTechCtrs.R}

\item{outputPath}{Path to save files, plots and the report}

\item{sdCutOff}{Defines the sd cutoff of technical replicates to be accounted in the analysis. Default is 0.2}

\item{assayInfo}{Path to assay info csv file}

\item{sampleInfo}{Path to sample info csv file}
}
\value{
Flags possible outliers based on sdCutOff, filters them out, recalculates Ct means and plots before and after Ct values.
}
\description{
This function flags technical outliers based on the Ct standard deviation from sdCutOff. As deviations in technical
replicates typically arise from pipetting error, we can exclude them from skewing the data points.
}
