% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta_CT.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numerical vector}

\item{na.rm}{boolean to ignore NAN}
}
\value{
vector
}
\description{
Calculate the geometric average
}
