% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importFiles.R
\name{importFiles}
\alias{importFiles}
\title{qPCR Import Files}
\usage{
importFiles(inputPath, skiprows = 25)
}
\arguments{
\item{inputPath}{Path to Assay info, Sample info and Results file from QuantStudio 5 384-well instrument}

\item{skiprows}{Number of rows to skip from the Results file. Default is 25}
}
\value{
Imports tables and merges them creating a final dataframe with all sample and target info.
}
\description{
This function imports assay info, sample info and results file from QuantStudio5 384-well instrument and
merges all tables into a single dataframe with all information.
}
