% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta_delta_CT.R
\name{mutate_DDCT_avg}
\alias{mutate_DDCT_avg}
\title{Mutate ddct average}
\usage{
mutate_DDCT_avg(col1, col2)
}
\arguments{
\item{col1}{col1}

\item{col2}{col2}
}
\value{
interp for mutate
}
\description{
Interp function to use in mutate call of calculate_ddct
}
