% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta_delta_CT.R
\name{mutate_DDCT_sem}
\alias{mutate_DDCT_sem}
\title{Mutate ddct sem}
\usage{
mutate_DDCT_sem(col1, col2, col3)
}
\arguments{
\item{col1}{col1}

\item{col2}{col2}

\item{col3}{col3}
}
\value{
interp for mutate
}
\description{
Interp function ot use in mutate cell of calculate_ddct
}
