% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTechCtrs.R
\name{plotTechCtrs}
\alias{plotTechCtrs}
\title{qPCR Plot Technical Controls}
\usage{
plotTechCtrs(df, techCtrs, outputPath)
}
\arguments{
\item{df}{Dataframe imported from importFiles.}

\item{techCtrs}{Path to Assay info, Sample info and Results file from QuantStudio 5 384-well instrument}

\item{outputPath}{Path to save files, plots and the report}
}
\value{
Plots the technical controls (e.g., water, RT-) and saves plots and table in outuput path.
}
\description{
This function imports assay info, sample info and results file from QuantStudio5 384-well instrument.
}
