% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPCRreport.R
\name{qPCRreport}
\alias{qPCRreport}
\title{qPCR Analysis Report}
\usage{
qPCRreport(
  hsk = "cdk2ap2",
  ctr = "pcas13d",
  sdCutOff = 0.2,
  inputPath,
  outputPath,
  techCtrs = defaultTechCtrs,
  allGroups = TRUE,
  skiprows = 25
)
}
\arguments{
\item{hsk}{Defines a character string with the name of your housekeeping gene used to normalize the qPCR data. Default is 'cdk2ap2'}

\item{ctr}{Defines a character string with the name of your control group used to normalize delta-delta-Cts. Default is 'cas13d'}

\item{sdCutOff}{Defines the sd cutoff of technical replicates to be accounted in the analysis. Default is 0.2}

\item{inputPath}{Path to Assay info, Sample info and Results file from QuantStudio 5 384-well instrument}

\item{outputPath}{Path to save files, plots and the report}

\item{techCtrs}{Character string that defines the names of your technical controls, e.g. 'water', 'RT-', etc. Default string is within 'R/defaultTechCtrs.R'}

\item{allGroups}{TRUE or FALSE to determine if plots will include all conditions or only conditions that match the target name. Useful when you name your groups from knockdowns based on the target gene name. Default is TRUE.}

\item{skiprows}{Number of rows to skip from the Results file. Default is 25}
}
\value{
Plots, tables and HTML report are saved in the output folder path.
}
\description{
This function process your qPCR data and combines all results and processing into an HTML report.
All middle steps are also saved in the output folder path.
}
