% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTechCtrs.R
\name{plotTechCtrs}
\alias{plotTechCtrs}
\title{qPCR Plot Technical Controls}
\usage{
plotTechCtrs(df, techCtrs = defaultTechCtrs, outputPath)
}
\arguments{
\item{df}{Dataframe imported from importFiles.}

\item{techCtrs}{Character string that defines the names of your technical controls, e.g. 'water', 'RT-', etc. Default string is within 'R/defaultTechCtrs.R'}

\item{outputPath}{Path to save files, plots and the report}
}
\value{
Plots the technical controls (e.g., water, RT-) and saves plots and table in outuput path.
}
\description{
This function imports assay info, sample info and results file from QuantStudio5 384-well instrument.
}
